Reorg crates (#194)
* Move ReadOnlyDB trait into backend


* Move types and models into substrate-archive;
Merge sql_block_builder and models, block_exec and tasks


* Remove substrate-archive-common


* Fix binary


* Apply review suggestion


* Make clippy happy


* Revert license header and module documentation


* Update binary Cargo.lock


* Apply review suggestion
