+++ bin/node-template-archive/src/main.rs
@@ -27,7 +27,7 @@ use substrate_archive_backend::SecondaryRocksDB;
 
 pub fn main() -> anyhow::Result<()> {
 	let config = config::Config::new()?;
-	substrate_archive::init_logger(config.cli().log_level, log::LevelFilter::Debug)?;
+	substrate_archive::logger::init(config.cli().log_level, log::LevelFilter::Debug)?;
 
 	let mut archive = ArchiveBuilder::<Block, RuntimeApi, Executor, SecondaryRocksDB>::default()
 		.chain_spec(Box::new(config.cli().chain_spec.clone()))
