+++ bin/polkadot-archive/Cargo.toml
@@ -5,10 +5,6 @@ authors = ["Andrew Plaza <aplaza@liquidthink.net>"]
 edition = "2018"
 
 [dependencies]
-polkadot-service = { git = "https://github.com/paritytech/polkadot", branch = "master" }
-sp-io = { git = "https://github.com/paritytech/substrate", branch = "master" }
-sc-chain-spec = { git = "https://github.com/paritytech/substrate", branch = "master" }
-
 anyhow = "1.0"
 clap = { version = "2.33.1", features = ["yaml", "suggestions", "color"] }
 ctrlc = { version = "3.1.5", features = ["termination"] }
@@ -16,6 +12,9 @@ log = "0.4"
 serde = "1.0"
 toml = "0.5"
 
+polkadot-service = { git = "https://github.com/paritytech/polkadot", branch = "master" }
+sc-chain-spec = { git = "https://github.com/paritytech/substrate", branch = "master" }
+sp-io = { git = "https://github.com/paritytech/substrate", branch = "master" }
+
 substrate-archive = { path = "../../substrate-archive", features = ["logging"] }
-substrate-archive-common = { path = "../../substrate-archive-common" }
 substrate-archive-backend = { path = "../../substrate-archive-backend" }
