+++ substrate-archive-backend/Cargo.toml
@@ -36,11 +36,3 @@ sp-storage = { git = "https://github.com/paritytech/substrate", branch = "master
 sp-trie = { git = "https://github.com/paritytech/substrate", branch = "master" }
 sp-version = {  git = "https://github.com/paritytech/substrate", branch = "master" }
 sp-wasm-interface = {  git = "https://github.com/paritytech/substrate", branch = "master" }
-
-# Workspace
-substrate-archive-common = { path = '../substrate-archive-common' }
-
-[features]
-default = ["logging"]
-logging = ["substrate-archive-common/logging"]
-test_rocksdb = []
