+++ substrate-archive-backend/src/read_only_backend/state_backend.rs
@@ -27,7 +27,7 @@ use sp_core::storage::ChildInfo;
 use sp_runtime::traits::{Block as BlockT, HashFor};
 use sp_state_machine::{StateMachineStats, TrieBackend, UsageInfo as StateUsageInfo};
 
-use substrate_archive_common::ReadOnlyDB;
+use crate::database::ReadOnlyDB;
 
 /// DB-backed patricia trie state, transaction type is an overlay of changes to commit.
 pub type DbState<B> = TrieBackend<Arc<dyn sp_state_machine::Storage<HashFor<B>>>, HashFor<B>>;
