--- substrate-archive-common/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "substrate-archive-common"
-version = "0.1.0"
-authors = ["Parity Technologies <admin@parity.io>"]
-edition = "2018"
-
-[dependencies]
-dirs = { version = "3", package = "directories" }
-log = "0.4"
-serde = { version = "1.0", features = ["derive"] }
-sqlx = { version = "0.5", default-features = false, features = ["postgres", "macros", "runtime-async-std-rustls", "migrate", "json", "offline"] }
-xtra = { version = "0.5.0-rc.1", features = ["with-smol-1"] }
-
-# Parity
-codec = { package = "parity-scale-codec", version = "1.3", default-features = false, features = ["derive", "full"] }
-
-# Substrate
-sp-runtime = { git = "https://github.com/paritytech/substrate", branch = "master" }
-sp-storage = { git = "https://github.com/paritytech/substrate", branch = "master" }
-
-# Optional
-fern = { version = "0.6", features = ["colored"], optional = true }
-chrono = { version = "0.4", optional = true }
-
-[features]
-default = ["logging"]
-logging = ["chrono", "fern"]
